Added a common prefix for api. Client rooting.
Changed rooting on client side to reflect the changes made in the
transition from flask_restful to flask_restplus in
2927a1587a5706f3c6a90a514f7d3765dd31146c

It will be more confortable to use the new system, which will use a
function to generate rooting instead of string concatenation.

Added a single api prefix, which will be very comfy with proxy.config.